+++ app/users/[userKey]/inbox/_services/inboxService/handlers/undo.ts
@@ -1,7 +1,7 @@
 import type { User } from "@prisma/client";
 
 import { apSchemaService } from "@/_shared/activitypub/apSchemaService";
-import { userFindService } from "@/_shared/service/user";
+import { userFindService } from "@/_shared/user/services/userFindService";
 import { prisma } from "@/_shared/utils/prisma";
 
 import type { InboxError } from "./errors";
